app/src/main/java/com/dicoding/picodiploma/loginwithanimation/view/main/MainViewModel.kt
package com.dicoding.picodiploma.loginwithanimation.view.main

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.asLiveData
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.dicoding.picodiploma.loginwithanimation.data.StoryRepository
import com.dicoding.picodiploma.loginwithanimation.data.api.ListStoryItem
import com.dicoding.picodiploma.loginwithanimation.data.pref.UserModel
import kotlinx.coroutines.flow.catch
import kotlinx.coroutines.launch

class MainViewModel(private val storyRepository: StoryRepository) : ViewModel() {

    private val _isLoading = MutableLiveData<Boolean>()
    val isLoading: LiveData<Boolean> = _isLoading

    private val _error = MutableLiveData<String>()
    val error: LiveData<String> get() = _error

    val stories: LiveData<PagingData<ListStoryItem>> = storyRepository.getStoriesPaging().cachedIn(viewModelScope)

    fun refreshStories() {
        storyRepository.refreshStories()
    }

    fun fetchStories() {
        _isLoading.value = true
        viewModelScope.launch {
            storyRepository.getStoriesPaging().cachedIn(viewModelScope)
        }
    }

    fun getSession(): LiveData<UserModel> {
        _isLoading.value = true
        return storyRepository.getSession().asLiveData()
    }

    fun logout() {
        viewModelScope.launch {
            storyRepository.logout()
        }
    }

}